wip: move ExtensionBroadcastChannel into separate file
